Router を介して HomeView を表示する
